% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Affairs}
\alias{Affairs}
\title{Fair's Extramartial Affairs Data.}
\format{
An object of class \code{data.frame} with 601 rows and 11 columns.
}
\source{
AER package
}
\usage{
Affairs
}
\description{
Cross-section data from a survey conducted by Psychology Today in 1969.
An event-column was added for easier use as \link[survival:Surv]{survival::Surv}-object.
}
\details{
The variables in the dataframe are
\itemize{
\item \code{affairs} \link{numeric} How often engaged in affairs during the past year?
\item \code{gender} \link{factor} indicating gender.
\item \code{age} \link{numeric} coding age in years.
\item \code{religiousness} \link{numeric} coding religiousness from 1=anti, to 5=very
\item ...
\item \code{event} \link{numeric} status-coding, 0: left censored, 1: event
\item \code{event2} \link{numeric} status coding for left- and right-censoring, 0: right-censored, 1: event, 2: left-censored
}
}
\keyword{datasets}
