% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cens.R
\name{estimHessian}
\alias{estimHessian}
\title{Estimate ('observed') Hessian matrix at MLE}
\usage{
estimHessian(opt, thetaParInd)
}
\arguments{
\item{opt}{optimization result}

\item{thetaParInd}{indeces for variance components (theta) in parameter vector}
}
\value{
Hessian matrix at MLE for fixed-effect coefficients
}
\description{
Estimate ('observed') Hessian matrix at MLE
}
