% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoring.R
\name{flattenResponse}
\alias{flattenResponse}
\title{Get an quick approximation for a flat response vector from the Surv-response.
It is useful to get sensible start-values using ordinary regression.}
\usage{
flattenResponse(yTime1, yTime2, yStat)
}
\description{
Get an quick approximation for a flat response vector from the Surv-response.
It is useful to get sensible start-values using ordinary regression.
}
