% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getOptfun}
\alias{getOptfun}
\title{Get the optimizer function and perform minimal checks
Internal utility, only used in \link{optwrapCens}}
\usage{
getOptfun(optimizer)
}
\arguments{
\item{optimizer}{character string ( = function name) \emph{or} function}
}
\value{
optimizer function
}
\description{
Get the optimizer function and perform minimal checks
Internal utility, only used in \link{optwrapCens}
}
\seealso{
\link[lme4:optwrap]{lme4::optwrap}
}
