% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical.R
\name{int_gh}
\alias{int_gh}
\title{Numeric integration with Gauss-Hermite quadrature.}
\usage{
int_gh(f, ord = 7, ...)
}
\arguments{
\item{f}{constant or function to be integrated after factoring out the weight function w = exp(-x^2), vectorized in its variable of integration}

\item{ord}{order of quadrature rule}

\item{...}{additional arguments passed down to integrand function f}
}
\value{
numeric, approximation of integral
}
\description{
Weight function is exp(-x^2) and it is an improper integral from -Inf to Inf.
E.g. for standard normal density integration, use f=1 constant and get sqrt(pi)-times the integral value.
}
