% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cens.R
\name{lmcens.objFun}
\alias{lmcens.objFun}
\title{Factory method to create the objective function for linear regression with censored observations.}
\usage{
lmcens.objFun(x, yTime1, yTime2, yStat, w)
}
\arguments{
\item{x}{model matrix nxp}

\item{yTime1}{first response time}

\item{yTime2}{second response time}

\item{yStat}{status variable in interval style}

\item{w}{weights vector}
}
\value{
negative log-likelihood as objective function

negative log-likelihood for given parameter vector for data sample, with positive likelihood contribution of individual observations as attribute
gradient of negative log-likelihood function
weights are on the log-likelihood scale

gradient numeric vector of length equal to number of parameters
}
\description{
It is the negative log-likelihood function. Analytical gradient is provided.
Optimization via \code{\link[stats:optim]{stats::optim()}}.
}
