% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cens.R
\name{lmcens}
\alias{lmcens}
\title{Fit a linear model with censored observations.}
\usage{
lmcens(formula, data, subset, weights, contrasts = NULL, offset,
  start = NULL, ...)
}
\description{
Residuals are not implemented, yet. What is it for censored observations?
}
\seealso{
\link[stats:lm]{stats::lm}
}
