% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4cens-package.R
\docType{package}
\name{lme4cens}
\alias{lme4cens}
\alias{lme4cens-package}
\title{Linear Mixed Effect Models that cope with censored data.
It uses the formula interface of lme4. As a response variable you can use \code{Surv}-objects from the \code{survival} package.}
\description{
Linear Mixed Effect Models that cope with censored data.
It uses the formula interface of lme4. As a response variable you can use \code{Surv}-objects from the \code{survival} package.
}
