% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cens.R
\name{lmercens_rInt}
\alias{lmercens_rInt}
\title{Linear mixed models with random intercept and with censored response.}
\usage{
lmercens_rInt(formula, data = NULL, REML = TRUE, control = lmerControl(),
  start = NULL, verbose = 0L, subset, weights, na.action, offset,
  contrasts = NULL, devFunOnly = FALSE, quadrature = c("gh", "stats"),
  ...)
}
\description{
This function is modelled after the official function \code{lmer}.
}
