% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{log1mexp}
\alias{log1mexp}
\title{Efficient calculation of \code{log(1-exp(x))}.}
\usage{
log1mexp(x)
}
\arguments{
\item{x}{numeric, non-positive values}
}
\description{
useful for some likelihood calculations, e.g. exponential distribution with left-censoring.
}
\seealso{
Martin Maechler
}
