% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical.R
\name{logxmy}
\alias{logxmy}
\title{Calculates \code{log(x-y)} where \code{x > y} holds}
\usage{
logxmy(lx, ly)
}
\arguments{
\item{lx}{1st argument on log-scale}

\item{ly}{2nd argument on log-scale}
}
\description{
Calculates \code{log(x-y)} where \code{x > y} holds
}
