% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{logxpy}
\alias{logxpy}
\title{Calculates log(x+y)
Avoid underflow when numbers are small.}
\usage{
logxpy(lx, ly)
}
\arguments{
\item{lx}{1st argument on log-scale}

\item{ly}{2nd argument on log-scale}
}
\description{
Calculates log(x+y)
Avoid underflow when numbers are small.
}
\seealso{
http://stackoverflow.com/questions/5802592/dealing-with-very-small-numbers-in-r
}
