% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimizeLmerCens}
\alias{optimizeLmerCens}
\title{Optimization function for linear mixed models with censoring.}
\usage{
optimizeLmerCens(
  devfun,
  optimizer = formals(lmerControl)$optimizer,
  restart_edge = formals(lmerControl)$restart_edge,
  boundary.tol = formals(lmerControl)$boundary.tol,
  start = NULL,
  verbose = 0L,
  control = list(),
  ...
)
}
\arguments{
\item{devfun}{a deviance function}
}
\value{
Results of an optimization.
}
\description{
Optimization function for linear mixed models with censoring.
}
\seealso{
\code{\link[lme4:modular]{lme4::optimizeLmer()}}
}
