% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optwrapCens}
\alias{optwrapCens}
\title{optwrap for censored responses.}
\usage{
optwrapCens(optimizer, fn, par, lower = -Inf, upper = Inf,
  control = list(), calc.derivs = TRUE, verbose = 0L)
}
\description{
Started as a copy of \link[lme4:::optwrap]{lme4::::optwrap}
then adapted to the censoring situaion (e.g. optimization is for all parameters) and simplified the code (removed some options).
}
\seealso{
\link[lme4:optwrap]{lme4::optwrap}
}
