% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cens.R
\name{refit.lmercens}
\alias{refit.lmercens}
\title{Refit a lmercens model with a different response.}
\usage{
\method{refit}{lmercens}(object, newresp, control = NULL,
  devFunOnly = NULL, start = NULL, verbose = NULL, ...)
}
\arguments{
\item{newresp}{a \link[survival:Surv]{survival::Surv}-object carrying a censored response}
}
\value{
a lmercens-object from the new censored respone variable
}
\description{
Basically, open up the existing lmercens-fit and inject the new response and call \link{lmercens} again.
}
