% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sigma.lmercens}
\alias{sigma.lmercens}
\title{Extract variance estimates on standard-deviation scale}
\usage{
\method{sigma}{lmercens}(object, which = c("residual", "between"), ...)
}
\arguments{
\item{which}{which type of variance estimate}
}
\description{
Extract variance estimates on standard-deviation scale
}
