% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleepstudy}
\alias{sleepstudy}
\title{Sleepstudy data with additional censoring information (event2).}
\format{An object of class \code{data.frame} with 180 rows and 6 columns.}
\source{
lme4 package
}
\usage{
sleepstudy
}
\description{
Sleepstudy data with additional censoring information (event2).
}
\keyword{datasets}
