% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cens.R
\name{vcov.lmcens}
\alias{vcov.lmcens}
\title{Variance covariance function for \link{lmcens}-models.}
\usage{
\method{vcov}{lmcens}(obj)
}
\value{
asymptotic variance-covariance matrix for the parameters, including residual std error as last parameter
}
\description{
Includes the residual standard error as own parameter
}
