% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{vcov.lmercens}
\alias{vcov.lmercens}
\title{Variance-covariance matrix for fixed effect coefficients.
It uses the observed Fisher information matrix at the ML parameter estimate.
Therefore, it is only asymptotically usefule.}
\usage{
\method{vcov}{lmercens}(obj, ...)
}
\arguments{
\item{obj}{a fitted \code{lmercens}-object}
}
\value{
estimated variance-covariance matrix for fixed effect coefficients
}
\description{
Variance-covariance matrix for fixed effect coefficients.
It uses the observed Fisher information matrix at the ML parameter estimate.
Therefore, it is only asymptotically usefule.
}
