% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cens.R
\name{lmcens.fit}
\alias{lmcens.fit}
\title{Internal workhorse function for linear regression with censored observations.}
\usage{
lmcens.fit(x, y, w, offset = NULL, tol = 1e-07)
}
\arguments{
\item{x}{model matrix}

\item{y}{response that uses the interval-type censoring}

\item{w}{weights vector}

\item{offset}{offset vector that is subtracted from the time variable (in
case of interval-censoring both boundaries are adapted)}
}
\value{
list with regression fit stuff (e.g. coefficients, fitted.values
effects, rank, ..)
negative log-likelihood
}
\description{
It uses maximum likelihood famework to come up with estimates.
Optimization via \code{\link[stats:optim]{stats::optim()}}.
}
