% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmod_cens.R
\name{lmcens.objFun}
\alias{lmcens.objFun}
\title{Objective function for linear regression with censored observations}
\usage{
lmcens.objFun(x, yTime1, yTime2, yStat, w, offset)
}
\arguments{
\item{x}{model matrix nxp}

\item{yTime1}{first response time}

\item{yTime2}{second response time}

\item{yStat}{status variable in interval style}

\item{w}{weights vector}

\item{offset}{offset vector, can be \code{NULL}}
}
\value{
negative log-likelihood as objective function
}
\description{
Factory method to create the negative log-likelihood function. The
(analytical) gradient is provided for improved performance for optimization.
}
