% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4cens-package.R
\docType{package}
\name{lme4cens}
\alias{lme4cens}
\title{Simple Linear Mixed Effect models that cope with censored data.
It uses the formula interface of \code{lme4}-package. As a response variable you can use \code{Surv}-objects from the \code{survival} package.}
