% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cens.R
\name{lmercens}
\alias{lmercens}
\title{Simple random intercept mixed models with censored response.}
\usage{
lmercens(formula, data = NULL, REML,
  control = lme4::lmerControl(optimizer = "optimx", optCtrl = list(method
  = "L-BFGS-B")), start = NULL, verbose = 0L, subset, weights,
  na.action, offset, contrasts = NULL, devFunOnly = FALSE, ...)
}
\arguments{
\item{control}{list-like control object of class \code{\link[lme4]{lmerControl}}. Defaults to use \code{optimx}'s BFGS optimizer.}
}
\value{
lmercens object
}
\description{
This function is modelled like the main function \code{\link[lme4]{lmer}}.
}
