% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{logxpy}
\alias{logxpy}
\title{Calculates \code{log(x+y)} where the (small positive) arguments are given on log-scale.}
\usage{
logxpy(lx, ly)
}
\arguments{
\item{lx}{1st argument on log-scale}

\item{ly}{2nd argument on log-scale}
}
\description{
This implementation avoids underflow when numbers are small. The trick is to do the summation on log-space.
The implementation is vectorized.
}
\seealso{
http://stackoverflow.com/questions/5802592/dealing-with-very-small-numbers-in-r
}
