% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cens.R
\name{mkLmerCensDevfun_rInt_R}
\alias{mkLmerCensDevfun_rInt_R}
\title{Factory for objective function closures for simple LMER-models with censored response.}
\usage{
mkLmerCensDevfun_rInt_R(
  fr,
  X,
  reTrms,
  REML = FALSE,
  verbose = 0,
  control,
  formula = stop("provide model formula"),
  start,
  ...
)
}
\arguments{
\item{fr}{model frame with response variable in its first column}

\item{X}{design matrix of fixed effects}

\item{reTrms}{list with random effect terms}

\item{control}{an object of class \code{lmerControl}}

\item{formula}{model formula passed from \code{lmerCens} to get start values}

\item{start}{list of start values}

\item{param}{parameter vector, first the variance parameters}
}
\value{
objective function which maps the parameters to the corresponding negative log-likelihood
}
\description{
This implements the objective function -- i.e. the negative log-likelihood -- for simple scalar random effect models with censoring.
It is implemented in R. The variance parameters and the fixed effect parameter are not profiled out of the log-likelihood
because of the censoring.
The variance parameters are on log-scale (as to avoid \emph{constrained} optimization).
It uses ML, REML currently not implemented.
}
