% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cens.R
\name{mkLmerCensDevfun_rInt_R}
\alias{mkLmerCensDevfun_rInt_R}
\title{Objective function closure for LMER-models with censored response.}
\usage{
mkLmerCensDevfun_rInt_R(fr, X, reTrms, REML = TRUE, start = NULL,
  verbose = 0, quadrature = c("stats", "gh"), ...)
}
\arguments{
\item{fr}{dataframe with response variable in its first column}

\item{reTrms}{list with random effect terms}
}
\value{
objective function which maps the covariance parameters theta to the corresponding restricted deviance
}
\description{
This implements the objective function for simple scalar models with censoring.
It is implemented in R. The variance parameter and the fixed effect parameter is not profiled out because of the censoring.
}
