% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cens.R
\name{mkLmerCensDevfun_rInt_R}
\alias{mkLmerCensDevfun_rInt_R}
\title{Factory for objective function closures for simple LMER-models with censored response.}
\usage{
mkLmerCensDevfun_rInt_R(fr, X, reTrms, REML = FALSE, verbose = 0, control,
  formula = stop("provide model formula"), ...)
}
\arguments{
\item{fr}{model frame with response variable in its first column}

\item{X}{design matrix of fixed effects}

\item{reTrms}{list with random effect terms}

\item{control}{an object of class \code{lmerControl}}

\item{formula}{model formula passed from \code{lmerCens} to get start values}

\item{param}{parameter vector, first the variance parameters}
}
\value{
objective function which maps the parameters to the corresponding negative log-likelihood
}
\description{
This implements the negative log-likelihood (=objective function) for simple scalar random effect models with censoring.
It is implemented in R. The variance parameter and the fixed effect parameter is not profiled out because of the censoring.
The variance parameters are on log-scale (as to avoid boundary issues at zero).
It uses ML, REML not implemented.
}
