% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleepstudy2}
\alias{sleepstudy2}
\title{Balanced subset of Sleepstudy with only four patients, including censoring information (event2).}
\format{An object of class \code{data.frame} with 40 rows and 5 columns.}
\source{
lme4 package
}
\usage{
sleepstudy2
}
\description{
Balanced subset of Sleepstudy with only four patients, including censoring information (event2).
}
\details{
For testing purposes with \code{interval}-type censoring status
\itemize{
\item \code{event2} \link{numeric} status variable
\item \code{event3} \link{numeric} status variable with clear-cut censoring boundaries (<312 L and > 350 R)
}
}
\keyword{datasets}
