% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleepstudy2}
\alias{sleepstudy2}
\title{Balanced subset of Sleepstudy with only four patients, including censoring information (event2)
For testing purposes.}
\format{An object of class \code{data.frame} with 40 rows and 4 columns.}
\source{
lme4 package
}
\usage{
sleepstudy2
}
\description{
Balanced subset of Sleepstudy with only four patients, including censoring information (event2)
For testing purposes.
}
\keyword{datasets}
